requirements.txt
tensorflow==2.10.0
numpy==1.21.6
Pillow==9.2.0